project/code_submission/requirements/requirements_gemini.txt
# Google Gemini 版本 RAG 系统依赖
# 适用于使用 Google Gemini API 的场景

# 核心 LangChain 包
langchain>=0.1.0
langchain-community>=0.0.20
langchain-chroma>=0.1.0
langchain-google-genai

# 向量数据库
chromadb>=0.4.0
sentence-transformers>=2.2.0

# Google AI SDK
google-generativeai
python-dotenv>=1.0.0

# 数据处理
numpy>=1.24.0
